// core/enums/supported_social_media_platforms.enum.ts
/* eslint-disable no-unused-vars */

export enum SupportedSocialMediaPlatforms {
    Facebook = 'Facebook',
    Twitter = 'Twitter',
    Instagram = 'Instagram',
    LinkedIn = 'LinkedIn',
    WhatsApp = 'WhatsApp',
    Viber = 'Viber',
    Zoom = 'Zoom',
    Discord = 'Discord',
    Slack = 'Slack',
    GitHub = 'GitHub'
}